move sidebar to a new file
